+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `ex.BoundingBox.translate(pos: Vector)` will create a new bounding box shifted by `pos`
 - `ex.BoundingBox.scale(scale: Vector)` will create a new bounding box scaled by `scale`
 - Added `isActor()` and `isCollider()` type guards
+- Added `ex.CollisionShape.draw` collision shapes can now be drawn, actor's will use these shapes if no other drawing is specified
 
 ### Changed